fix(textblock): [Android] Fix TextBlock color not updating
Certain paths, such as a combination of infinite-width containers and calling the UpdateLayout() method, could result in parts of the visual tree falling into a bad layout state where IsLayoutRequested=true is interleaved with IsLayoutRequested=false, resulting in the bad subtrees never updating. To prevent this, check for this state and repair it by setting the layout request on the parent of the affected view.

This ensures that TextBlock will update correctly when its Foreground color changes (since it relies on a measure+arrange pass to be redrawn).